Refactor dataset document download helpers
