+++ api/agent-notes/controllers/console/datasets/datasets_document.py.md
@@ -16,8 +16,7 @@
 
   - Only supported when `Document.data_source_type == "upload_file"`.
   - Performs dataset permission + tenant checks via `DocumentResource.get_document(...)`.
-  - Validates the referenced `UploadFile` exists for the same tenant.
-  - Returns a signed URL generated by `core.file.helpers.get_signed_file_url(..., as_attachment=True)`.
+  - Delegates `Document -> UploadFile` validation and signed URL generation to `DocumentService.get_document_download_url(...)`.
   - Applies `cloud_edition_billing_rate_limit_check("knowledge")` to match other KB operations.
   - Response body is **only**: `{ "url": "<signed-url>" }`.
 
@@ -27,8 +26,8 @@
   - Returns `application/zip` as a single attachment download.
   - Rationale: browsers often block multiple automatic downloads; a ZIP avoids that limitation.
   - Applies `cloud_edition_billing_rate_limit_check("knowledge")`.
-  - The attachment filename (`download_name`) is a random `.zip` value (client-visible only).
-  - Uses batch queries to load documents and upload files before streaming the ZIP.
+  - Delegates dataset permission checks, document/upload-file validation, and download-name generation to
+    `DocumentService.prepare_document_batch_download_zip(...)` before streaming the ZIP.
 
 ## Verification plan
 
@@ -41,10 +40,11 @@
 
 ## Shared helper
 
-- `_get_upload_file_id_for_upload_file_document(...)` normalizes the upload file id and enforces source checks.
-- `_get_upload_file_for_upload_file_document(document)` centralizes the “Document → UploadFile” lookup.
-- `_get_upload_files_by_document_id_for_zip_download(...)` batches lookups and returns a
-  `document_id -> UploadFile` map for ZIP downloads.
+- `DocumentService.get_document_download_url(document)` resolves the `UploadFile` and signs a download URL.
+- `DocumentService.prepare_document_batch_download_zip(...)` performs dataset permission checks, batches
+  document + upload file lookups, preserves request order, and generates the client-visible ZIP filename.
+- Internal helpers now live in `DocumentService` (`_get_upload_file_id_for_upload_file_document(...)`,
+  `_get_upload_file_for_upload_file_document(...)`, `_get_upload_files_by_document_id_for_zip_download(...)`).
 - ZIP packing is handled by `FileService.build_upload_files_zip_tempfile(...)`, which also:
   - sanitizes entry names to avoid path traversal, and
   - deduplicates names while preserving extensions (e.g., `doc.txt` → `doc (1).txt`).
