+++ api/agent-notes/services/dataset_service.py.md
@@ -8,6 +8,8 @@
 - Tenant checks must be enforced consistently across dataset/document operations.
 - `DocumentService.get_documents_by_ids(...)` fetches documents for a dataset using `id.in_(...)`.
 - `FileService.get_upload_files_by_ids(...)` performs tenant-scoped batch lookup for `UploadFile` (dedupes ids with `set(...)`).
+- `DocumentService.get_document_download_url(...)` and `prepare_document_batch_download_zip(...)` handle
+  dataset/document permission checks plus `Document -> UploadFile` validation for download endpoints.
 
 ## Verification plan
