+++ api/agent-notes/tests/unit_tests/controllers/console/datasets/test_datasets_document_download.py.md
@@ -13,7 +13,7 @@ Unit tests for the console dataset document download endpoint:
   - `DocumentService.get_document` for single-file download tests
   - `DocumentService.get_documents_by_ids` + `FileService.get_upload_files_by_ids` for ZIP download tests
   - `FileService.get_upload_files_by_ids` for `UploadFile` lookups in single-file tests
-  - `core.file.helpers.get_signed_file_url` to return a deterministic URL
+  - `services.dataset_service.file_helpers.get_signed_file_url` to return a deterministic URL
 - Document mocks include `id` fields so batch lookups can map documents by id.
 
 ## Covered cases
